pyproject.toml
[tool.poetry]
name = "assigment-2"
version = "0.1.0"
description = ""
authors = ["Aziza <azizamukashkyzy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-debug-toolbar = "^5.0.1"
drf-spectacular = "^0.28.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
fake = "^0.8"
factory-boy = "^3.3.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
coverage = "^7.6.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "--nomigrations -v -s --disable-warnings"
python_files = ["test_*.py"]
pythonpath = [
    "."
]